Change fees and amounts on transactions from strings to numbers

+++ service/src/main/kotlin/io/provenance/explorer/config/Properties.kt
@@ -27,7 +27,7 @@ class ExplorerProperties {
 
     fun isMetricJobEnabled() = txsMetricJobEnabled.toBoolean()
 
-    fun minGasPrice() = minimumGasPrice.toDouble()
+    fun minGasPrice() = minimumGasPrice.toBigDecimal()
 
 }
